+++ lib/crsdialer/crsdialer.go
@@ -6,21 +6,18 @@ package crsdialer
  * Easy dialer to connect to curlrevshell
  * By J. Stuart McMurray
  * Created 20250905
- * Last Modified 20250924
+ * Last Modified 20260815
  */
 
 import (
 	"context"
-	"crypto/sha256"
 	"crypto/tls"
-	"encoding/base64"
 	"errors"
 	"fmt"
 	"io"
 	"net"
 	"net/http"
 	"os"
-	"strings"
 	"syscall"
 
 	"github.com/magisterquis/curlrevshell/lib/sstls"
@@ -28,17 +25,15 @@ import (
 
 // SHA256Prefix is allowed but not required to be at the start of the
 // fingerprint passed to Dial.
-const SHA256Prefix = "sha256//"
+const SHA256Prefix = sstls.SHA256Prefix
 
 // ErrNoMatchingCertificate indicates a TLS connection's peer did not present
 // a certificate matching a configured fingerprint.
-var ErrNoMatchingCertificate = errors.New(
-	"no certificate with correct fingerprint found",
-)
+var ErrNoMatchingCertificate = sstls.ErrNoMatchingCertificate
 
 // Dial connects to curlrevshell at the given server HTTPS URL.  The server's
 // TLS certificate is expected to have the given fingerprint as printed by
-// curlrevshell, though it allowed to omit SHA256Prefix.
+// curlrevshell, though it allowed to omit [sstls.SHA256Prefix].
 // The returned os.File, which is really a socketpair, is proxied to and from
 // the HTTPS connection, which will be closed when the os.File is closed or the
 // context is done.
@@ -123,49 +118,15 @@ func Dial(ctx context.Context, serverURL, fingerprint string) (
 	return cConn, nil
 }
 
-// TLSFingerprintVerifier returns a function which can be used for
-// [tls.Config.VerifyConnection].  It ensures the peer presents a certificate
-// with the given fingerprint, which must be a base64-encoded sha256 hash as
-// used by curl, with or without the leading sha256//.
+// Deprecated: As of curlrevshell v0.0.1-beta.9, this function simply calls
+// [sstls.TLSFingerprintVerifier].
+//
+//go:fix inline
 func TLSFingerprintVerifier(wantFP string) (
 	func(tls.ConnectionState) error,
 	error,
 ) {
-	/* Don't need the prefix. */
-	wantFP = strings.TrimPrefix(wantFP, SHA256Prefix)
-	/* Make sure the fingerprint looks correct. */
-	if b, err := base64.StdEncoding.DecodeString(wantFP); nil != err {
-		return nil, fmt.Errorf("decoding fingerprint: %w", err)
-	} else if sha256.Size != len(b) {
-		return nil, fmt.Errorf(
-			"decoded fingerprint not %d bytes",
-			sha256.Size,
-		)
-	}
-
-	/* Return a function to check if any of the certs in
-	cs.PeerCertificates have the right hash. */
-	return func(cs tls.ConnectionState) error {
-		/* Check ALL the certs. */
-		for i, cert := range cs.PeerCertificates {
-			/* Hash the server's cert. */
-			gotFP, err := sstls.PubkeyFingerprint(cert)
-			if nil != err {
-				return fmt.Errorf(
-					"getting fingerprint for server "+
-						"certificate %d/%d: %w",
-					i+1, len(cs.PeerCertificates),
-					err,
-				)
-			}
-
-			/* See if it matches. */
-			if gotFP == wantFP {
-				return nil
-			}
-		}
-		return ErrNoMatchingCertificate
-	}, nil
+	return sstls.TLSFingerprintVerifier(wantFP)
 }
 
 // unixConn turns fd into a net.UnixConn.