+++ lib/sstls/archive.go
@@ -5,12 +5,11 @@ package sstls
  * Read and Save certs with an archive file
  * By J. Stuart McMurray
  * Created 20240327
- * Last Modified 20261011
+ * Last Modified 20260815
  */
 
 import (
 	"crypto/tls"
-	"crypto/x509"
 	"errors"
 	"fmt"
 	"os"
@@ -21,15 +20,10 @@ import (
 	"golang.org/x/tools/txtar"
 )
 
-// ErrCacheFileEmpty indicates the the file passed to LoadCachedCertificate was
-// empty.
-var ErrCacheFileEmpty = errors.New("cache file empty")
-
 // LoadCachedCertificate loads the certificate from the named file, which
 // should have been created with SaveCertificate.
 func LoadCachedCertificate(certFile string) (tls.Certificate, error) {
-	/* Read the saved cert. */
-	ta, err := txtar.ParseFile(certFile)
+	b, err := os.ReadFile(certFile)
 	if nil != err {
 		return tls.Certificate{}, fmt.Errorf(
 			"reading %s: %w",
@@ -37,6 +31,14 @@ func LoadCachedCertificate(certFile string) (tls.Certificate, error) {
 			err,
 		)
 	}
+	return loadCachedCertificate(certFile, b)
+}
+
+// loadCachedCertificate does what LoadCachedCertificate says it does but
+// from a pre-loaded file.  certFile is only used in error messages.
+func loadCachedCertificate(certFile string, b []byte) (tls.Certificate, error) {
+	/* Read the saved cert. */
+	ta := txtar.Parse(b)
 
 	/* Grab the important files. */
 	var certB, keyB []byte
@@ -66,15 +68,12 @@ func LoadCachedCertificate(certFile string) (tls.Certificate, error) {
 		)
 	}
 
-	/* Make sure Leaf is set. */
-	leaf, err := x509.ParseCertificate(cert.Certificate[0])
-	if nil != err {
-		return tls.Certificate{}, fmt.Errorf(
-			"parsing read leaf: %w",
-			err,
-		)
+	/* Make sure Leaf is set.  At one point, tls.X509KeyPair didn't
+	always do this, but it should now unless someone set
+	GODEBUG=x509keypairleaf=0. */
+	if nil == cert.Leaf {
+		return tls.Certificate{}, ErrLeafCertificateNotSet
 	}
-	cert.Leaf = leaf
 
 	return cert, nil
 }